+++ ocs_ci/helpers/dr_helpers.py
@@ -1,6 +1,7 @@
 """
 Helper functions specific for DR
 """
+
 import json
 import logging
 import tempfile
@@ -1098,258 +1099,3 @@ def verify_drpolicy_cli(switch_ctx=None):
         raise UnexpectedBehaviour(
             f"DRPolicy is not in succeeded or validated state: {status}"
         )
-
-
-def enable_cluster_backup():
-    """
-    set cluster-backup to True in mch resource
-    Note: changing this flag automatically installs OADP operator
-    """
-    mch_resource = ocp.OCP(
-        kind=constants.ACM_MULTICLUSTER_HUB,
-        resource_name=constants.ACM_MULTICLUSTER_RESOURCE,
-        namespace=constants.ACM_HUB_NAMESPACE,
-    )
-    mch_resource._has_phase = True
-    resource_dict = mch_resource.get()
-    for components in resource_dict["spec"]["overrides"]["components"]:
-        if components["name"] == "cluster-backup":
-            components["enabled"] = True
-    mch_resource_yaml = tempfile.NamedTemporaryFile(
-        mode="w+", prefix="mch", delete=False
-    )
-    yaml_serialized = yaml.dump(resource_dict)
-    mch_resource_yaml.write(yaml_serialized)
-    mch_resource_yaml.flush()
-    run_cmd(f"oc apply -f {mch_resource_yaml.name}")
-    mch_resource.wait_for_phase("Running")
-    backup_pod_status_check()
-
-
-def create_s3_bucket(access_key, secret_key, bucket_name):
-    """
-    Create s3 bucket
-    Args:
-        access_key (str): S3 access key
-        secret_key (str): S3 secret key
-        acm_indexes (list): List of acm indexes
-    """
-    client = boto3.resource(
-        "s3",
-        verify=True,
-        endpoint_url="https://s3.amazonaws.com",
-        aws_access_key_id=access_key,
-        aws_secret_access_key=secret_key,
-    )
-    try:
-        client.create_bucket(
-            Bucket=bucket_name,
-            CreateBucketConfiguration={"LocationConstraint": constants.AWS_REGION},
-        )
-        logger.info(f"Successfully created backup bucket: {bucket_name}")
-    except BotoCoreError as e:
-        logger.error(f"Failed to create s3 bucket {e}")
-        raise
-
-
-def build_bucket_name(acm_indexes):
-    """
-    Create backupname from cluster names
-    Args:
-        acm_indexes (list): List of acm indexes
-    """
-    bucket_name = ""
-    for index in acm_indexes:
-        bucket_name += config.clusters[index].ENV_DATA["cluster_name"]
-    return bucket_name
-
-
-@retry((TimeoutExpiredError, ACMClusterConfigurationException), tries=20, delay=10)
-def backup_pod_status_check():
-    pods_list = get_all_pods(namespace=constants.ACM_HUB_BACKUP_NAMESPACE)
-    if len(pods_list) != 3:
-        raise ACMClusterConfigurationException("backup pod count mismatch ")
-    for pod in pods_list:
-        # check pod status Running
-        if not pod.data["status"]["phase"] == "Running":
-            raise ACMClusterConfigurationException("backup pods not in Running state")
-
-
-def create_generic_credentials(access_key, secret_key, acm_indexes):
-    """
-    Create s3 secret for backup and restore
-    Args:
-        access_key (str): S3 access key
-        secret_key (str): S3 secret key
-        acm_indexes (list): List of acm indexes
-    """
-    s3_cred_str = (
-        "[default]\n"
-        f"aws_access_key_id={access_key}\n"
-        f"aws_secret_access_key={secret_key}\n"
-    )
-    cred_file = tempfile.NamedTemporaryFile(mode="w+", prefix="s3_creds", delete=False)
-    cred_file.write(s3_cred_str)
-    cred_file.flush()
-
-    cmd = (
-        f"oc create secret generic cloud-credentials --namespace {constants.ACM_HUB_BACKUP_NAMESPACE} "
-        f"--from-file cloud={cred_file.name}"
-    )
-    old_index = config.cur_index
-    # Create on all ACM clusters
-    for index in acm_indexes:
-        config.switch_ctx(index)
-        try:
-            run_cmd(f"oc create namespace {constants.ACM_HUB_BACKUP_NAMESPACE}")
-        except CommandFailed as ex:
-            if "already exists" in str(ex):
-                logger.warning("Namespace already exists!")
-            else:
-                raise
-        try:
-            run_cmd(cmd)
-        except CommandFailed:
-            logger.error("Failed to create generic secrets cloud-credentials")
-    config.switch_ctx(old_index)
-
-
-def enable_managed_serviceaccount():
-    """
-    update MultiClusterEngine
-
-    """
-    old_ctx = config.cur_index
-    config.switch_ctx(get_active_acm_index())
-
-    multicluster_engine = ocp.OCP(
-        kind="MultiClusterEngine",
-        resource_name=constants.MULTICLUSTER_ENGINE,
-    )
-    multicluster_engine._has_phase = True
-    resource = multicluster_engine.get()
-    for item in resource["spec"]["overrides"]["components"]:
-        if item["name"] == "managedserviceaccount":
-            item["enabled"] = True
-    multicluster_engine_yaml = tempfile.NamedTemporaryFile(
-        mode="w+", prefix="multiengine", delete=False
-    )
-    yaml_serialized = yaml.dump(resource)
-    multicluster_engine_yaml.write(yaml_serialized)
-    multicluster_engine_yaml.flush()
-    run_cmd(f"oc apply -f {multicluster_engine_yaml.name}")
-    multicluster_engine.wait_for_phase("Available")
-    config.switch_ctx(old_ctx)
-
-
-def create_dpa(bucket_name):
-    """
-    create DPA
-    OADP will be already installed when we enable backup flag
-    Here we will create dataprotection application and
-    update bucket name and s3 storage link
-    Args:
-        bucket_name (str): Name of the Bucket
-    """
-    oadp_data = templating.load_yaml(constants.ACM_DPA)
-    oadp_data["spec"]["backupLocations"][0]["velero"]["objectStorage"][
-        "bucket"
-    ] = bucket_name
-    oadp_yaml = tempfile.NamedTemporaryFile(mode="w+", prefix="oadp", delete=False)
-    templating.dump_data_to_temp_yaml(oadp_data, oadp_yaml.name)
-    run_cmd(f"oc create -f {oadp_yaml.name}")
-    # Validation
-    validate_dpa()
-
-
-@retry((CommandFailed, ACMClusterConfigurationException), tries=10, delay=10)
-def validate_dpa():
-    """
-    Validate
-    1. 3 restic / Node pods
-    2. 1 velero pod
-    3. backupstoragelocation resource in "Available" phase
-    """
-    # Check restic pods.
-    # Restic pods have been renamed to node-agent after oadp 1.2
-    oadp_version = get_oadp_version()
-
-    if version.compare_versions(f"{oadp_version} >= 1.2"):
-        restic_or_node_pod_prefix = "node-agent"
-    else:
-        restic_or_node_pod_prefix = "restic"
-    restic_or_node_list = get_pods_having_label(
-        f"name={restic_or_node_pod_prefix}", constants.ACM_HUB_BACKUP_NAMESPACE
-    )
-    if len(restic_or_node_list) != constants.RESTIC_OR_NODE_POD_COUNT:
-        raise ACMClusterConfigurationException("restic/node pod count mismatch")
-    for pod in restic_or_node_list:
-        if pod["status"]["phase"] != "Running":
-            raise ACMClusterConfigurationException(
-                "restic/node pod not in 'Running' phase"
-            )
-
-    # Check velero pod
-    veleropod = get_pods_having_label(
-        "app.kubernetes.io/name=velero", constants.ACM_HUB_BACKUP_NAMESPACE
-    )
-    if len(veleropod) != constants.VELERO_POD_COUNT:
-        raise ACMClusterConfigurationException("Velero pod count mismatch")
-    if veleropod[0]["status"]["phase"] != "Running":
-        raise ACMClusterConfigurationException("Velero pod not in 'Running' phase")
-
-    # Check backupstoragelocation resource in "Available" phase
-    backupstorage = ocp.OCP(
-        kind="BackupStorageLocation",
-        resource_name="default",
-        namespace=constants.ACM_HUB_BACKUP_NAMESPACE,
-    )
-    resource = backupstorage.get()
-    if resource["status"].get("phase") != "Available":
-        raise ACMClusterConfigurationException(
-            "Backupstoragelocation resource is not in 'Avaialble' phase"
-        )
-    logger.info("Dataprotection application successful")
-
-
-def validate_secret_creation_oadp():
-    """
-    Verify Secret are created
-
-    Raises:
-        ResourceNotFoundError: raised when secret not found
-
-    """
-    try:
-        secret = ocp.OCP(
-            kind=constants.SECRET,
-            resource_name="cloud-credentials",
-            namespace=constants.ACM_HUB_BACKUP_NAMESPACE,
-        )
-        secret.get()
-        logger.info("Secret found")
-    except CommandFailed:
-        raise ResourceNotFoundError("Secret Not found")
-
-
-def validate_policy_compliance_status(
-    resource_name, resource_namespace, compliance_state
-):
-    """
-    Validate policy status for given resource
-
-    Raises:
-        ResourceWrongStatusException: Raised when resource state does not match
-
-    """
-
-    compliance_output = ocp.OCP(
-        kind=constants.ACM_POLICY,
-        resource_name=resource_name,
-        namespace=resource_namespace,
-    )
-    compliance_status = compliance_output.get()
-    if compliance_status["status"]["compliant"] == compliance_state:
-        logger.info("Compliance status Matches ")
-    else:
-        raise ResourceWrongStatusException("Compliance status does not match")
